Move all mysql client code out of server
